+++ clients/client-sdk/src/index.ts
@@ -2,4 +2,6 @@ export { IdentityClient } from './clients/identity';
 
 export { ChannelClient } from './clients/channel';
 
+export { ApiVersion } from './models/apiVersion';
+
 export * from './models';